cloud-functions/video-processor/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.16.0
pymongo==4.6.0
requests==2.31.0
ffmpeg-python==0.2.0
